refactor: Model_NFW to submodule ModelNFW
This refactor is to enable better unit testing. We cannot (as far as I
understand) access functions within functions for testing (i.e
gnfw_gas_radial_term). Furthermore, if we just refactor this flat in the
file, this will be added to the global namespace, which is unnecessary
pollution for anything bar testing.
If the file becomes a submodule, then we can still only export
Model_NFW, but now can access ModelNFW.gnfw_gas_radial_term (after it is
refactored flat in the file) for testing.
Note that we need to do a little bit more import/include at the start of
the ModelNFW file as the namespace is not inherited from parent module
BayesJ.
We do have to also update cluster_models.jl to also use and export the
model for completeness.