refactor: Allow non-integer values in obs or bg sample
In the case where pre-processing of observations are performed which
results in non-integer counts in the background or observation we should
be able to process these.
Need to update the typing for functions as the functions themselves can
already handle floats. As a part of this, sample can now take observed
and observed_background which may have different types.